feat: Add notification screen and localization keys
This commit adds a new notification screen to the app and also adds the necessary localization keys for it. It also includes some minor changes such as updating the elevation of a dropdown button and fixing a typo in a file name. Additionally, it adds a loading state for the search screen while the necessary data is being fetched.